fix: Permission issues with XDG_RUNTIME_DIR
This makes it so that the parent directories are created recursively
with 0755 permissions, but still owned by root.

Previously, when the service is running as root and the parent
directories to XDG_RUNTIME_DIR do not yet exist they are created with
0700 permissions and owned by root. This results in issues with the
compositor being unable to interact with XDG_RUNTIME_DIR even though it
exists.

Signed-off-by: Randolph Sapp <rs@ti.com>