feat(login): trust the counter more than the mtime
This trusts the counter value unless the mtime is older than 2 seconds.
Previously we only trusted the counter if the mtime was within 2
seconds.

This is more applicable for machines without an RTC where system time
between reboots may confuse the previous login rate limit mechanism.

We still allow the mtime to override the counter before we check the
retry count specifically to handle AutologinMaxRetry events. We hope
that some mechanism has held the process for at least 2 seconds before
we continue.

Signed-off-by: Randolph Sapp <rs@ti.com>