handleInterrupt: handle unruly applications
Emptty cannot currently exit via an external signal if the session
process exits abnormally. This conflicts with the expected behavior of
systemd, which expects that the service will always respect signals.

This has the added benefit of allowing users to use ^C to restart the
login process at any point in the menu system.

Signed-off-by: Randolph Sapp <rs@ti.com>